made the project more java-like
