+++ com/studmane/nlpserver/handler/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * http handlers for incoming API requests
+ */
+package com.studmane.nlpserver.handler;
