+++ com/studmane/nlpserver/service/exception/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains exceptions thrown by services
+ */
+package com.studmane.nlpserver.service.exception;
