+++ com/studmane/nlpserver/service/model/MessageIntent.java
@@ -1,0 +1,12 @@
+package com.studmane.nlpserver.service.model;
+
+/**
+ * A Message Intent declares the intent of the message
+ */
+public enum MessageIntent {
+    NEGATIVE, AFFIRMATIVE,
+    SCHEDULE, RESCHEDULE,
+    CANCEL,
+    QUERY_WHEN,
+    OTHER
+}
