+++ com/studmane/nlpserver/service/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Service objects handle overall functionality of the API
+ * This is where most of the crazy NLP logic goes down
+ */
+package com.studmane.nlpserver.service;
