+++ com/studmane/nlpserver/service/response/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Serializable response objects returned from the API services
+ */
+package com.studmane.nlpserver.service.response;
