--- src/handler/package-info.java
@@ -1,4 +1,0 @@
-/**
- * http handlers for incoming API requests
- */
-package handler;
