--- src/service/exception/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains exceptions thrown by services
- */
-package service.exception;
