--- src/service/model/MessageIntent.java
@@ -1,12 +1,0 @@
-package service.model;
-
-/**
- * A Message Intent declares the intent of the message
- */
-public enum MessageIntent {
-    NEGATIVE, AFFIRMATIVE,
-    SCHEDULE, RESCHEDULE,
-    CANCEL,
-    QUERY_WHEN,
-    OTHER
-}
