--- src/service/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Service objects handle overall functionality of the API
- * This is where most of the crazy NLP logic goes down
- */
-package service;
