--- src/service/request/package-info.java
@@ -1,8 +1,0 @@
-/**
- * Request objects, used for passing data to the API service objects
- * 
- * AnalyisRequest is used to to make a request to the AnalysisService. It contains
- * a message to be analyzed as well as a conversation and/or metadata about the
- * recipients to aid in processing.
- */
-package service.request;
