--- src/service/response/Serializable.java
@@ -1,13 +1,0 @@
-package service.response;
-/** Serializable Interface
- * 
- * Serializable objects can be serialized to JSON strings
- */
-public interface Serializable {
-
-    /** Serializes the object
-     * 
-     * @return a JSON string containing the object
-     */
-    public String Serialize();
-};
