--- src/service/response/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Serializable response objects returned from the API services
- */
-package service.response;
